requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-jinja2==1.6
aiosignal==1.3.2
arrow==1.3.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.5
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
numpy==2.2.3
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-gitlab==5.6.0
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
six==1.17.0
textual==1.0.0
textual-dev==1.7.0
textual-serve==1.1.1
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
uc-micro-py==1.0.3
urllib3==2.3.0
yarl==1.18.3
